requirements.txt
feedparser==6.0.11
arxiv==2.2.0
google-generativeai==0.8.3
resend==2.13.0
jinja2==3.1.4
python-dateutil==2.9.0
pytz==2024.2
pyyaml==6.0.2
python-dotenv==1.0.1
tenacity==9.0.0
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.3
pdf2image==1.17.0
Pillow==10.4.0